+++ react/components/IframeContainer.tsx
@@ -1,10 +1,12 @@
 import type { ReactNode } from 'react'
 import React from 'react'
-import { ThemeProvider } from '@vtex/admin-ui'
+import { createSystem, ThemeProvider } from '@vtex/admin-ui'
 
 import { IframeLegacy } from './IframeLegacy'
 import { Iframe } from './Iframe'
 import { DELOREAN_REGISTRY } from '../util'
+
+const system = createSystem('iframe-container')
 
 const isLegacy = (slug: string) =>
   String(slug).startsWith('Site/') ||
@@ -21,14 +23,18 @@ export function IframeContainer(props: Props) {
 
   if (withoutLegacy) {
     // Cover just IO apps, ignore legacy apps.
-    return <Iframe params={params} />
+    return (
+      <ThemeProvider system={system}>
+        <Iframe params={params} />
+      </ThemeProvider>
+    )
   }
 
   if (isLegacy(slug)) {
     // IframeLegacy and isLegacy are covering Catalog and Legacy CMS admins
     // those are the old portal9 admins served by portal in vtexcommercestable
     return (
-      <ThemeProvider>
+      <ThemeProvider system={system}>
         <IframeLegacy params={params} />
       </ThemeProvider>
     )
@@ -37,13 +43,21 @@ export function IframeContainer(props: Props) {
   if (isDeloreanAdmin(slug)) {
     // This is covering "Legacy" admins versioned by Delorean
     // those were rendered by concierge, but now render delivers them as statics
-    return <Iframe params={params} isDeloreanAdmin />
+    return (
+      <ThemeProvider system={system}>
+        <Iframe params={params} isDeloreanAdmin />
+      </ThemeProvider>
+    )
   }
 
   // This covers the rest of cases, which are IO apps that are admins
   // with "/admin/app" routes and fallback to 404 slugs
   // The customHeightGap is only used on the Admin V4 (see vtex/admin)
-  return <Iframe params={params} customHeightGap={customHeightGap} />
+  return (
+    <ThemeProvider system={system}>
+      <Iframe params={params} customHeightGap={customHeightGap} />
+    </ThemeProvider>
+  )
 }
 
 interface Props {