+++ react/core/Navbar/index.tsx
@@ -1,16 +1,16 @@
 import type { ReactNode } from 'react'
 import type { FlexProps, StyleProp } from '@vtex/admin-ui'
-import { Heading, jsxs, Flex, merge, Button } from '@vtex/admin-ui'
-import { IconArrow } from '@vtex/admin-ui-icons'
+import {
+  createComponent,
+  Heading,
+  jsxs,
+  Flex,
+  merge,
+  Button,
+  IconArrow,
+} from '@vtex/admin-ui'
 
-export function Navbar(props: NavbarProps) {
-  const navbarProps = useNavbar(props)
-
-  return jsxs({
-    component: Flex,
-    props: navbarProps,
-  })
-}
+export const Navbar = createComponent(Flex, useNavbar)
 
 export function useNavbar(props: NavbarProps): FlexProps {
   const { styleOverrides, children, title, link, ...rest } = props
@@ -24,30 +24,24 @@ export function useNavbar(props: NavbarProps): FlexProps {
     styles: merge(theme, styleOverrides),
     children: [
       link
-        ? jsxs({
-            component: Button,
-            props: {
-              variant: 'tertiary',
-              onClick: () => link.onClick,
-              icon: jsxs({
-                component: IconArrow,
-                props: {
-                  direction: 'left',
-                  title: link.label,
-                },
-              }),
-            },
+        ? jsxs(Button, {
+            variant: 'tertiary',
+            onClick: () => link.onClick,
+            icon: jsxs(IconArrow, {
+              direction: 'left',
+              title: link.label,
+            }),
           })
         : null,
-      jsxs({
-        component: Heading,
-        props: {
+      jsxs(
+        Heading,
+        {
           styleOverrides: {
             fontSettings: 'medium',
           },
         },
-        children: title,
-      }),
+        title
+      ),
     ],
     align: 'center',
     ...rest,